Point the tab guards at the tab that is there now
Deleting the Credits fixture left two robot examples naming it. One is a build
error the workspace never sees: `robot_tab_screenshot_dump` lists
`DemoTab::CranorbitCredits` among the tabs whose captures cannot be compared
across runs, and robot examples need `robot-app`, so neither `cargo test
--workspace` nor `clippy --all-targets` builds them. It failed the GPU robot
job at compile time. The reason the entry exists survives the rename intact —
`WearState::advance` scrolls forever exactly as `CreditsState::advance` did, so
a capture after a fixed settle lands wherever the scroll happened to be — so
the entry moves to `DemoTab::Wear` rather than going away.

Behind that compile error sat a second failure it hid: `robot_credits_watch_fps`
looks its tab up by the label `"Credits (watch)"`, which no longer exists, and
panicked with `tab not found` the moment the build was fixed.

Repointing it would have been the wrong repair. It was written to catch glyph
metrics cached against the size they were measured at, and it could only do
that because the page under it re-measured several hundred glyphs a frame at a
size a curve moved. A scaling list scales a row through a layer transform; no
font size moves on this page any more, so that defect would not show here and
the test would have passed while guarding nothing. It is now
`robot_wear_watch_fps`, and its documentation says what it measures — the
heaviest text page against a page with almost no text — and says plainly that
the animated-size regression is guarded by
`a_font_size_animation_measures_each_glyph_once_rather_than_once_per_size` in
`cranpose-render-common` instead.

The bound is re-measured rather than reused. The old page ran 2x the reference
with `recomps=0`, because a `Canvas` on a frame clock recomposes nothing. This
one runs **7.7x** (0.495ms against 0.065ms) with **~1845 recompositions per
second against ~1877 frames** — the whole non-virtualised list recomposing every
frame while it scrolls. The bound is 12x: above what the page costs today with
room for host noise, so it still catches a text pipeline or a per-row layer tree
that starts costing a multiple of what it costs now. The recomposition-per-frame
cost itself is a property of the widget set, not something for this guard to
encode, and is raised on the PR.

Also fixed: the failure path called `std::process::exit(1)` from the driver
thread, which races the main thread's GPU teardown and dumps core — the run
above reported `exit=139`, which is how a real failure gets buried. It sets a
flag and returns an `ExitCode` from `main`, like the other runners.
